Make Install Less Invasive
Hundreds of modules will install the python-pip package, and have requirements
on it.  This module uses Package rather than enure_packages() which causes
puppet to complain about aliasing and possibly re-definition.  To play more
nicely with the rest of the eco-system use ensure_packages and make the
package name that of the actual package and not an alias.